Add Go-E charger integration and refactor SolarMan API code
In this commit, Go-E charger integration was added to the application. This provides the ability for users to set their charging speed directly from the user interface. It involved the introduction of new UI components, backend methods, and environment variables. Meanwhile, the SolarMan API code was refactored, migrating some logic to the SolarManApi class and simplifying the GET method in the api/suggest route.