Update app version and improve car charging status management
The app version has been updated in the package.json file. New functionality has been introduced to better manage the car charging status in the goERealTimeCard.tsx file. This function determines whether a car is waiting, charging, already charged, or in an unknown state. This status is then used to control the enabling of certain buttons. The charging behavior has also been adjusted in the status route where charging will not be initiated if the car is already charged.