Updated SolarMan API integration and incremented version
Expanded the API interaction with SolarMan to include getting station data. Improved the URL formatting when force flag is present in the 'realTime' endpoint. An attribute for stationId is now included in the result object when the API response is unsuccessful. Also updated the package version in package.json accordingly.