Add station list request and update application version
Added a POST request to get the list of stations in the `http/solarman.http` file. The response from this request is used to set a global `stationId` variable. The version in `package.json` is incremented to reflect these changes. Additionally, if a result is unsuccessful in `solarManApi.ts`, the URL endpoint is added to the result for easier debugging.