Update package version and refactor SolarMan API
The package version in package.json has been updated to 0.7.3. In addition, refactoring was done to the solarManApi.ts file. The API call logic was extracted into a separate function (getRealTimeInfoResponse) to avoid code duplication and to make the code more readable. This also makes the code easier to debug or extend in the future.